# Providers lock parallelized by platform