Provider locking is quite a slow process, with binaries needing to be downloaded for every architecture that you expect to run tofu on.  This change fetches and computes checksums for binaries from each platform in parallel.

In conjunction with https://github.com/opentofu/opentofu/pull/2729, every provider version platform is downloaded in their own routine.  This can both speed up the process dramatically, or potentially get you rate limited.

Anecdata:
* 5 large providers on two architectures: 25s -> 20s
* 5 large providers on two architectures: 38s -> 28s

Does not resolve, but helps with https://github.com/opentofu/opentofu/issues/1442

## Target Release

1.10.0

## Checklist

<!-- Please check of ALL items in this list for all PRs: -->

- [x] I have read the [contribution guide](https://github.com/opentofu/opentofu/blob/main/CONTRIBUTING.md).
- [x] I have not used an AI coding assistant to create this PR.
- [x] I have written all code in this PR myself OR I have marked all code I have not written myself (including modified code, e.g. copied from other places and then modified) with a comment indicating where it came from.
- [x] I (and other contributors to this PR) have not looked at the Terraform source code while implementing this PR.

### Go checklist

<!-- If your PR contains Go code, please make sure you check off all items on this list: --> 

- [x] I have run golangci-lint on my change and receive no errors relevant to my code.
- [x] I have run existing tests to ensure my code doesn't break anything.
- [x] I have added tests for all relevant use cases of my code, and those tests are passing.
- [x] I have only exported functions, variables and structs that should be used from other packages.
- [x] I have added meaningful comments to all exported functions, variables, and structs.

### Website/documentation checklist

<!-- If you have changed the website, please follow this checklist: -->

- [ ] I have locally started the website as [described here](https://github.com/opentofu/opentofu/blob/main/website/README.md) and checked my changes.