@lblazewski-whisker 2025-06-27
Looking forward for this feature as indeed there is a significant difference when running terraform provider lock and tofu provider lock.