@cam72cam 2025-07-08
I've been testing this a bit more and it seems that github can throttle you when a few parallel download requests are made.